+++ data/src/main/kotlin/dev/esbi/mizan/data/local/entity/CategorySpendingEntity.kt
@@ -3,12 +3,14 @@ package dev.esbi.mizan.data.local.entity
 import androidx.room.Entity
 import androidx.room.PrimaryKey
 
+import java.math.BigDecimal
+
 @Entity(tableName = "category_spending")
 data class CategorySpendingEntity(
     @PrimaryKey val category: String,
     val categoryLabel: String,
-    val totalAmount: Double,
+    val totalAmount: BigDecimal,
     val transactionCount: Int,
-    val percentage: Double,
+    val percentage: BigDecimal,
     val colorToken: String
 )
