+++ presentation/src/main/kotlin/dev/esbi/mizan/presentation/feature/dashboard/domain/model/DashboardSummary.kt
@@ -1,11 +1,12 @@
 package dev.esbi.mizan.presentation.feature.dashboard.domain.model
 
 import dev.esbi.mizan.domain.model.Transaction
+import java.math.BigDecimal
 
 data class DashboardSummary(
-    val totalBalance: Double,
-    val monthlyExpenses: Double,
-    val monthlySavings: Double,
+    val totalBalance: BigDecimal,
+    val monthlyExpenses: BigDecimal,
+    val monthlySavings: BigDecimal,
     val budgetLimit: Double,
     val budgetPercentageUsed: Double,
     val topCategories: List<CategorySpending>,
